Add files via upload
Is is now possible to set single threads active. This way the user don't has to generate the full wordlist at once if he doesn't has f.x. the storage. Also this gives the opportunity to use multiple computers for one wordlist. It is basically a preview of something that could be used later maybe even for a generator network.